requirements.txt
attrs
requests
tomli